pyproject.toml
[project]
name = "voice-core"
version = "0.1.0"
description = "Domain-neutral acoustic analysis and real-time voice processing engine"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "scipy",
    "librosa",
    "soundfile",
    "praat-parselmouth",
    "torchcrepe",
    "torch",
    "torchaudio",
    "pyworld",
    "praatio",
    "sounddevice",
    "setuptools>=82.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_core"]
